Rename and add some classes
The mecanum and sewerve kinematics have been rename to adding a drive. Also added to new scripts that will ne used for the odometry. The swerve and mecanum subsystems are now on a folder name subsystems. This ones now have functionality. The hardwareMap has been added to constants to facilitate the acces to this Class. The kinematics classes now has the normalize function as static allowing to use them without stablishing an instance.